+++ kernel/dev/serial.c
@@ -42,6 +42,9 @@ void serial_write_char(char c) {
 }
 
 void serial_write_string(const char *str) {
+    if (str == 0) {
+        str = "(null)";
+    }
     while (*str) {
         serial_write_char(*str++);
     }